Refactored database
