+++ scripts/database.js
@@ -43,40 +43,5 @@ export const database = {
         }
     ]
 }
-export const addCustomOrder = () => {
-    //Copy the current state of user choices
-    const newOrder = {...database.orderBuilder}
-
-    //add a new primary key to the object
-    const lastIndex = database.customOrders.length - 1
-    newOrder.id = database.customOrders[lastIndex].id + 1
-
-    //add a timestamp to the order
-    newOrder.timestamp = Date.now()
-
-    //add the new order object to custom orders state
-    database.customOrders.push(newOrder)
-
-    //reset the temporary state for user choices
-    database.orderBuilder = {}
-
-    //Broadcast a notification that permanent state has changed
-    document.dispatchEvent(new CustomEvent("stateChanged"))
-}
 
 
-export const getSizes = () => {
-    return database.sizes.map(size => ({...size}))
-}
-export const getOptions = () => {
-    return database.options.map(option => ({...option}))
-}
-
-
-export const setSize = (id) => {
-    database.orderBuilder.sizeId = id
-}
-
-export const setOption = (id) => {
-    database.orderBuilder.optionId = id
-}
